Change Exception to NonFoundException extends RuntimeException. Delete Interfece ProductService. Rename ProductServiceImpl to ProductService.
