+++ src/main/java/com/kodilla/ecommercee/exception/NotFoundException.java
@@ -1,0 +1,7 @@
+package com.kodilla.ecommercee.exception;
+
+public class NotFoundException extends RuntimeException {
+    public NotFoundException(String message) {
+        super(message);
+    }
+}
