--- src/main/java/com/kodilla/ecommercee/exception/ProductNotFoundException.java
@@ -1,7 +1,0 @@
-package com.kodilla.ecommercee.exception;
-
-public class ProductNotFoundException extends Throwable {
-    public ProductNotFoundException() {
-        super("Product not found");
-    }
-}
